Fixed changes in run_batch so that we now use parallel and have the right indexing in the match term.
Added more text to the Rmd.  Was able to cleanly run the file but there are excess messages and some errors in the processing that make the html look a bit junky.

Updated the settings file so that the clean_run is now true.

Added the `add_accumulation` function.  I think I'd like to, more broadly, make an `add_element` function, but I'm not sure how complicated that would be.